Update to new Xeryon version (#6)
* use new Xeryon version and add tcp connection feature with minimal changes to oem libraries

* Update README.md

Minor updates to ensure users know to load all 3 axes and so that the axis definitions match the axis definitions in HISPEC (A, B, C instead of X, Y, Z) and the stage definitions match the HISPEC stages (XLS_5_3N).

* Update README.md

Add comment showing users how to point to a specific settings file.

* changes after CR, cleanup

---------

Co-authored-by: Daniel Echeverri <47041464+danechever@users.noreply.github.com>